+++ src/Services/ResourceService.php
@@ -47,8 +47,7 @@ use UnzerSDK\Resources\PaymentTypes\Giropay;
 use UnzerSDK\Resources\PaymentTypes\HirePurchaseDirectDebit;
 use UnzerSDK\Resources\PaymentTypes\Ideal;
 use UnzerSDK\Resources\PaymentTypes\Invoice;
-use UnzerSDK\Resources\PaymentTypes\InvoiceFactoring;
-use UnzerSDK\Resources\PaymentTypes\InvoiceGuaranteed;
+use UnzerSDK\Resources\PaymentTypes\InvoiceSecured;
 use UnzerSDK\Resources\PaymentTypes\Paypal;
 use UnzerSDK\Resources\PaymentTypes\PIS;
 use UnzerSDK\Resources\PaymentTypes\Prepayment;
@@ -513,11 +512,23 @@ class ResourceService implements ResourceServiceInterface
     {
         $resourceType = IdService::getResourceTypeFromIdString($typeId);
         switch ($resourceType) {
+            case IdStrings::ALIPAY:
+                $paymentType = new Alipay();
+                break;
+            case IdStrings::BANCONTACT:
+                $paymentType = new Bancontact();
+                break;
             case IdStrings::CARD:
                 $paymentType = new Card(null, null);
                 break;
+            case IdStrings::EPS:
+                $paymentType = new EPS();
+                break;
             case IdStrings::GIROPAY:
                 $paymentType = new Giropay();
+                break;
+            case IdStrings::HIRE_PURCHASE_DIRECT_DEBIT:
+                $paymentType = new HirePurchaseDirectDebit();
                 break;
             case IdStrings::IDEAL:
                 $paymentType = new Ideal();
@@ -525,11 +536,16 @@ class ResourceService implements ResourceServiceInterface
             case IdStrings::INVOICE:
                 $paymentType = new Invoice();
                 break;
+            case IdStrings::INVOICE_FACTORING:
             case IdStrings::INVOICE_GUARANTEED:
-                $paymentType = new InvoiceGuaranteed();
+            case IdStrings::INVOICE_SECURED:
+                $paymentType = new InvoiceSecured();
                 break;
             case IdStrings::PAYPAL:
                 $paymentType = new Paypal();
+                break;
+            case IdStrings::PIS:
+                $paymentType = new PIS();
                 break;
             case IdStrings::PREPAYMENT:
                 $paymentType = new Prepayment();
@@ -537,35 +553,17 @@ class ResourceService implements ResourceServiceInterface
             case IdStrings::PRZELEWY24:
                 $paymentType = new Przelewy24();
                 break;
-            case IdStrings::SEPA_DIRECT_DEBIT_GUARANTEED:
-                $paymentType = new SepaDirectDebitGuaranteed(null);
-                break;
             case IdStrings::SEPA_DIRECT_DEBIT:
                 $paymentType = new SepaDirectDebit(null);
+                break;
+            case IdStrings::SEPA_DIRECT_DEBIT_GUARANTEED:
+                $paymentType = new SepaDirectDebitGuaranteed(null);
                 break;
             case IdStrings::SOFORT:
                 $paymentType = new Sofort();
                 break;
-            case IdStrings::PIS:
-                $paymentType = new PIS();
-                break;
-            case IdStrings::EPS:
-                $paymentType = new EPS();
-                break;
-            case IdStrings::ALIPAY:
-                $paymentType = new Alipay();
-                break;
             case IdStrings::WECHATPAY:
                 $paymentType = new Wechatpay();
-                break;
-            case IdStrings::INVOICE_FACTORING:
-                $paymentType = new InvoiceFactoring();
-                break;
-            case IdStrings::HIRE_PURCHASE_DIRECT_DEBIT:
-                $paymentType = new HirePurchaseDirectDebit();
-                break;
-            case IdStrings::BANCONTACT:
-                $paymentType = new Bancontact();
                 break;
             default:
                 throw new RuntimeException('Invalid payment type!');